Break command functions out into separate files
